test(screenspace): split test_screenspace.py into tests/screenspace/
Break the 4,257-line file (51 classes, 243 tests) into 11 themed files under
tests/screenspace/ grouped by analysis concern. Shared frame/region builders
move to tests/screenspace/_ss_helpers.py (bare-module import, matching existing
tests/ precedent). Pure move: no test logic changed, same 243 tests pass.

Left fake_scan stubs inline — they differ in signature and body and only share
a name, so consolidating would be a leaky abstraction rather than real dedup.
