+++ tests/screenspace/test_change_similarity.py
@@ -1,0 +1,72 @@
+"""Tests for frame diff, region similarity, phash, and scene fingerprint."""
+
+import numpy as np
+
+import screenspace
+
+
+class TestComputeFrameDiff:
+    def test_identical_frames(self):
+        frame = np.random.randint(0, 255, (50, 50, 3), dtype=np.uint8)
+        diff = screenspace.compute_frame_diff(frame, frame.copy())
+        assert diff == 0.0
+
+    def test_completely_different(self):
+        black = np.zeros((50, 50, 3), dtype=np.uint8)
+        white = np.full((50, 50, 3), 255, dtype=np.uint8)
+        diff = screenspace.compute_frame_diff(black, white)
+        assert diff > 0.9
+
+
+class TestRegionsAreSimilar:
+    def test_identical(self):
+        frame = np.random.randint(50, 200, (50, 50, 3), dtype=np.uint8)
+        is_similar, score = screenspace.regions_are_similar(frame, frame.copy())
+        assert is_similar is True
+        assert score >= 0.99
+
+    def test_different(self):
+        a = np.zeros((50, 50, 3), dtype=np.uint8)
+        b = np.full((50, 50, 3), 255, dtype=np.uint8)
+        is_similar, score = screenspace.regions_are_similar(a, b)
+        assert is_similar is False
+        assert score < 0.5
+
+
+class TestComputePhash:
+    def test_deterministic(self):
+        region = np.random.randint(0, 255, (64, 64, 3), dtype=np.uint8)
+        hash1 = screenspace.compute_phash(region)
+        hash2 = screenspace.compute_phash(region.copy())
+        assert hash1 == hash2
+
+    def test_different_images(self):
+        a = np.zeros((64, 64, 3), dtype=np.uint8)
+        b = np.full((64, 64, 3), 255, dtype=np.uint8)
+        hash_a = screenspace.compute_phash(a)
+        hash_b = screenspace.compute_phash(b)
+        assert hash_a != hash_b
+
+
+class TestSceneFingerprint:
+    def test_same_frame_similar(self):
+        frame = np.random.randint(50, 200, (50, 50, 3), dtype=np.uint8)
+        fp1 = screenspace.compute_scene_fingerprint(frame)
+        fp2 = screenspace.compute_scene_fingerprint(frame.copy())
+        score = screenspace.compare_scene_fingerprints(fp1, fp2)
+        assert score >= 0.99
+
+    def test_different_frames_dissimilar(self):
+        a = np.zeros((50, 50, 3), dtype=np.uint8)
+        b = np.full((50, 50, 3), 255, dtype=np.uint8)
+        fp_a = screenspace.compute_scene_fingerprint(a)
+        fp_b = screenspace.compute_scene_fingerprint(b)
+        score = screenspace.compare_scene_fingerprints(fp_a, fp_b)
+        assert score < 0.8
+
+    def test_fingerprint_has_expected_keys(self):
+        frame = np.random.randint(0, 255, (30, 30, 3), dtype=np.uint8)
+        fp = screenspace.compute_scene_fingerprint(frame)
+        assert "histogram" in fp
+        assert "edge_density" in fp
+        assert "color_stats" in fp
