+++ tests/screenspace/test_template.py
@@ -1,0 +1,179 @@
+"""Tests for template matching primitives and heatmap."""
+
+import numpy as np
+
+import screenspace
+from _ss_helpers import _make_icon, _make_icon_frame
+
+
+class TestMatchTemplate:
+    def test_exact_match(self):
+        # Use a textured pattern so template matching works after blur
+        rng = np.random.RandomState(42)
+        frame = rng.randint(0, 255, (100, 200, 3), dtype=np.uint8)
+        template = frame[30:60, 80:140].copy()
+        results = screenspace.match_template(frame, template, threshold=0.9)
+        assert len(results) >= 1
+        assert results[0]["score"] >= 0.9
+
+    def test_no_match(self):
+        frame = np.zeros((100, 200, 3), dtype=np.uint8)
+        template = np.full((20, 40, 3), 128, dtype=np.uint8)
+        results = screenspace.match_template(frame, template, threshold=0.9)
+        assert len(results) == 0
+
+    def test_template_larger_than_frame(self):
+        frame = np.zeros((20, 20, 3), dtype=np.uint8)
+        template = np.zeros((50, 50, 3), dtype=np.uint8)
+        results = screenspace.match_template(frame, template, threshold=0.5)
+        assert results == []
+
+    def test_match_with_mask(self):
+        rng = np.random.RandomState(42)
+        frame = rng.randint(0, 255, (100, 200, 3), dtype=np.uint8)
+        template = frame[30:60, 80:140].copy()
+        # Full opaque mask — should behave like no mask
+        mask = np.full((30, 60), 255, dtype=np.uint8)
+        results = screenspace.match_template(frame, template, threshold=0.9, mask=mask)
+        assert len(results) >= 1
+
+    def test_match_with_none_mask(self):
+        rng = np.random.RandomState(42)
+        frame = rng.randint(0, 255, (100, 200, 3), dtype=np.uint8)
+        template = frame[30:60, 80:140].copy()
+        results = screenspace.match_template(frame, template, threshold=0.9, mask=None)
+        assert len(results) >= 1
+
+
+class TestPrepareTemplateMask:
+    def test_binarizes_alpha_mask(self):
+        """Mask should come out as strictly 0 or 255 (no soft-blurred edges)."""
+        template = np.full((30, 60, 3), 128, dtype=np.uint8)
+        # Alpha ramp from 0..255 to exercise the boundary
+        mask = np.zeros((30, 60), dtype=np.uint8)
+        for c in range(60):
+            mask[:, c] = int(c * 255 / 59)
+        _, gray_mask, _ = screenspace._prepare_template(template, mask)
+        assert gray_mask is not None
+        unique = set(np.unique(gray_mask).tolist())
+        assert unique.issubset({0, 255})
+
+    def test_none_mask_stays_none(self):
+        template = np.full((30, 60, 3), 128, dtype=np.uint8)
+        _, gray_mask, _ = screenspace._prepare_template(template, None)
+        assert gray_mask is None
+
+
+class TestScanTemplateControls:
+    """Cover template_scale addition to scan_template."""
+
+    def _patch_single_frame(self, monkeypatch, frame: np.ndarray) -> None:
+        def fake_scan(video_path, interval, callback, **kwargs):
+            callback(0.0, frame)
+
+        monkeypatch.setattr(screenspace, "scan_video_full_frames", fake_scan)
+        monkeypatch.setattr(screenspace, "_probe_video_meta", lambda p: (30.0, 1.0))
+
+    def test_scale_fixes_size_mismatch(self, monkeypatch):
+        """A 40px template should miss a 20px in-frame icon at scale 1.0
+        but hit at scale 0.5."""
+        frame = _make_icon_frame(400, 200, [(100, 50, 20)])
+        # Template at the original (larger) size — mimics an uploaded PNG
+        # captured at 2x the in-video rendering.
+        template = _make_icon(40)
+        self._patch_single_frame(monkeypatch, frame)
+
+        full_size = screenspace.scan_template(
+            "/fake.mp4",
+            {"x": 0, "y": 0, "w": 400, "h": 200},
+            template,
+            threshold=0.70,
+            template_scale=1.0,
+        )
+        assert full_size == []
+
+        scaled = screenspace.scan_template(
+            "/fake.mp4",
+            {"x": 0, "y": 0, "w": 400, "h": 200},
+            template,
+            threshold=0.70,
+            template_scale=0.5,
+        )
+        assert len(scaled) == 1
+        match = scaled[0]["matches"][0]
+        assert abs(match["x"] - 100) <= 2
+        assert abs(match["y"] - 50) <= 2
+
+    def test_transparent_mask_no_false_positives_on_blank_frame(self, monkeypatch):
+        """Mostly-transparent PNG + blank frame should yield no matches at
+        the default threshold now that the mask is binarized."""
+        template = np.full((32, 32, 3), 220, dtype=np.uint8)
+        mask = np.zeros((32, 32), dtype=np.uint8)
+        # Only a small opaque cross in the center
+        mask[14:18, :] = 255
+        mask[:, 14:18] = 255
+        blank = np.full((200, 300, 3), 30, dtype=np.uint8)
+        self._patch_single_frame(monkeypatch, blank)
+
+        results = screenspace.scan_template(
+            "/fake.mp4",
+            {"x": 0, "y": 0, "w": 300, "h": 200},
+            template,
+            threshold=0.70,
+            template_mask=mask,
+        )
+        assert results == []
+
+    def test_scaled_masked_template_no_explosion(self, monkeypatch):
+        """Regression: masked matching at non-1.0 scale must not report
+        thousands of matches. TM_CCOEFF_NORMED with sparse masks at
+        reduced scale previously produced near-1.0 scores at every
+        position."""
+        # Mostly-transparent 50x50 PNG with a small opaque central cross.
+        template = np.full((50, 50, 3), 220, dtype=np.uint8)
+        mask = np.zeros((50, 50), dtype=np.uint8)
+        mask[22:28, :] = 255
+        mask[:, 22:28] = 255
+        # Random frame, not containing the template.
+        rng = np.random.RandomState(3)
+        frame = rng.randint(0, 255, (360, 640, 3), dtype=np.uint8)
+        self._patch_single_frame(monkeypatch, frame)
+
+        results = screenspace.scan_template(
+            "/fake.mp4",
+            {"x": 0, "y": 0, "w": 640, "h": 360},
+            template,
+            threshold=0.70,
+            template_mask=mask,
+            template_scale=0.75,
+        )
+        total = sum(r["match_count"] for r in results)
+        assert total < 50, f"Expected few/no matches, got {total}"
+
+
+class TestGenerateTemplateHeatmap:
+    def test_basic_heatmap(self, tmp_path):
+        results = [
+            {
+                "timestamp": 1.0,
+                "matches": [{"x": 10, "y": 10, "w": 50, "h": 50, "score": 0.9}],
+            },
+            {
+                "timestamp": 2.0,
+                "matches": [{"x": 20, "y": 20, "w": 50, "h": 50, "score": 0.8}],
+            },
+        ]
+        out = str(tmp_path / "heatmap.png")
+        path = screenspace.generate_template_heatmap(results, 200, 200, out)
+        assert path == out
+        assert (tmp_path / "heatmap.png").is_file()
+        assert (tmp_path / "heatmap.png").stat().st_size > 0
+
+    def test_empty_results_returns_none(self, tmp_path):
+        out = str(tmp_path / "heatmap.png")
+        assert screenspace.generate_template_heatmap([], 200, 200, out) is None
+
+    def test_no_matches_returns_none(self, tmp_path):
+        results = [{"timestamp": 1.0, "matches": []}]
+        out = str(tmp_path / "heatmap.png")
+        assert screenspace.generate_template_heatmap(results, 200, 200, out) is None
